Add files via upload
There is a riddle: "What is this: blue, big, with a mustache and completely filled with hares?" Answer: a trolleybus. Let's imagine that you are asking this riddle to a user. Create a program that will read the answer from the console.
Problem statement:
The user has 3 attempts. After three answers, the program should terminate;
If the user enters "Trolleybus", we output "Correct!" to the console and exit the loop;
If the user enters "I give up", we output "Correct answer: trolleybus." to the console and exit the loop;
If the user enters any other answer, we output "Think again." to the console and continue the loop.